Significantly speed up multithreaded reads by changing spdk_wait_completions locking pattern.
Previously I was locking every qpair before processing IO. This was to
ensure that all waiting IO would be complete, and there would be no
pending IO, by the end of the function. However, this is unnecessary as
we only need the calling thread's IO to be complete by the end of the
function. Now, we lock one qpair, process all of its completions, then
unlock it and go to the next thread. Other threads may add IO to qpairs
after the calling thread runs completions for it, but the calling thread
won't care.